langgraph-example/my_agent/requirements.txt
langgraph>=0.0.15
langchain_anthropic>=0.1.0
tavily-python>=0.2.0
langchain_community>=0.0.10
langchain_openai>=0.0.2
streamlit>=1.31.0
lxml>=4.9.3
elasticsearch>=8.11.0
rank_bm25>=0.2.2
mecab-python>=1.0.6
graphviz>=0.20.1
python-dotenv>=1.0.0
